requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
scikit-learn
wordcloud
Pillow
requests
